requirements.txt
Django==5.2.7
djangorestframework==3.16.1
pandas==2.2.3
openpyxl==3.1.5
psycopg2-binary==2.9.9
